requirements.txt
SciPy
Numpy
PyGame
Pyglet
Matplotlib
Pillow
multipledispatch
